+++ src/Schema.php
@@ -411,7 +411,7 @@ final class Schema extends AbstractPdoSchema
      */
     private function loadColumnSchema(array $info): ColumnSchemaInterface
     {
-        $columnFactory = $this->db->getColumnBuilderClass()::columnFactory();
+        $columnFactory = $this->db->getColumnFactory();
 
         $dbType = $info['type'];
         /** @psalm-var ColumnInfoArray $info */