+++ tests/ConnectionTest.php
@@ -11,7 +11,7 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\IntegrityException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Mysql\Column\ColumnBuilder;
+use Yiisoft\Db\Mysql\Column\ColumnFactory;
 use Yiisoft\Db\Mysql\Tests\Support\TestTrait;
 use Yiisoft\Db\Tests\Common\CommonConnectionTest;
 use Yiisoft\Db\Transaction\TransactionInterface;
@@ -158,10 +158,10 @@ final class ConnectionTest extends CommonConnectionTest
         $db->createCommand('SELECT 1')->queryScalar();
     }
 
-    public function testGetColumnBuilderClass(): void
+    public function testGetColumnFactory(): void
     {
         $db = $this->getConnection();
 
-        $this->assertSame(ColumnBuilder::class, $db->getColumnBuilderClass());
+        $this->assertInstanceOf(ColumnFactory::class, $db->getColumnFactory());
     }
 }