+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Enh #347, #353: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
 - Bug #349, #352: Restore connection if closed by connection timeout (@Tigrov)
 - Enh #354: Separate column type constants (@Tigrov)
+- Enh #355: Realize `ColumnBuilder` class (@Tigrov)
 
 ## 1.2.0 March 21, 2024